+++ packages/api-extractor/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- fix default export declaration
 
 ## 1.2.0 - 2021/11/9